stm32: unified parts of stmclk implementations
- unified dbp, hsi, and lfclk functions
- all of them now work for all STM32 CPUs
- renamed from bdp to dbp as this is the expression used
  in the reference manuals (old naming was a typo...)
