+++ cpu/stm32f1/stmclk.c
@@ -154,56 +154,3 @@ void stmclk_init_sysclk(void)
     /* re-enable IRQs */
     irq_restore(is);
 }
-
-void stmclk_enable_hsi(void)
-{
-    RCC->CR |= RCC_CR_HSION;
-    while (!(RCC->CR & RCC_CR_HSIRDY)) {}
-}
-
-void stmclk_disable_hsi(void)
-{
-    /* we only disable the HSI clock if not used as input for the PLL and if
-     * not used directly as system clock */
-#ifdef CLOCK_HSE
-    if ((RCC->CFGR & RCC_CFGR_SWS) != SWSHSI) {
-        RCC->CR &= ~(RCC_CR_HSION);
-    }
-#endif
-}
-
-void stmclk_enable_lfclk(void)
-{
-#if CLOCK_LSE
-    stmclk_bdp_unlock();
-    RCC->BDCR |= RCC_BDCR_LSEON;
-    while (!(RCC->BDCR & RCC_BDCR_LSERDY)) {}
-    stmclk_bdp_lock();
-#else
-    RCC->CSR |= RCC_CSR_LSION;
-    while (!(RCC->CSR & RCC_CSR_LSIRDY)) {}
-#endif
-}
-
-void stmclk_disable_lfclk(void)
-{
-#if CLOCK_LSE
-    stmclk_bdp_unlock();
-    RCC->BDCR &= ~(RCC_BDCR_LSEON);
-    stmclk_bdp_lock();
-#else
-    RCC->CSR &= ~(RCC_CSR_LSION);
-#endif
-}
-
-void stmclk_bdp_unlock(void)
-{
-    periph_clk_en(APB1, RCC_APB1ENR_PWREN);
-    PWR->CR |= PWR_CR_DBP;
-}
-
-void stmclk_bdp_lock(void)
-{
-    PWR->CR &= ~(PWR_CR_DBP);
-    periph_clk_dis(APB1, RCC_APB1ENR_PWREN);
-}
